requirements-dev.txt
# Dépendances de développement
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Dépendances optionnelles pour les connecteurs
psycopg2-binary>=2.9.0  # PostgreSQL
boto3>=1.26.0          # AWS S3

python-dotenv